feat: prepare for deployment
