--- .github/workflows/publish.yml
@@ -1,33 +1,0 @@
-name: Publish
-
-on:
-  release:
-    types: [prereleased, released]
-
-jobs:
-  build:
-
-    runs-on: macos-latest
-
-    steps:
-      - uses: actions/checkout@v2
-      - name: Set up JDK
-        uses: actions/setup-java@v2
-        with:
-          distribution: 'zulu'
-          java-version: '11'
-      - name: Update version property
-        uses: Reedyuk/write-properties@v1.0.3
-        with:
-          path: 'gradle.properties'
-          property: 'MODULE_VERSION_NUMBER'
-          value: ${{ github.ref_name }}
-      - name: Publish
-        uses: eskatos/gradle-command-action@v1
-        with:
-          arguments: :publish
-        env:
-          sonatypeUsername: ${{ secrets.SONATYPEUSERNAME }}
-          sonatypePassword: ${{ secrets.SONATYPEPASSWORD }}
-          ORG_GRADLE_PROJECT_signingKey: ${{ secrets.GPG_PRIVATE_KEY }}
-          ORG_GRADLE_PROJECT_signingPassword: ${{ secrets.PASSPHRASE }}
