+++ gradle.properties
@@ -2,31 +2,38 @@ org.gradle.jvmargs=-Xmx2048m -Dfile.encoding=UTF-8
 android.useAndroidX=true
 kotlin.code.style=official
 
-POM_DEVELOPER_ID=GitLiveApp
-POM_DEVELOPER_NAME=Andrew Reed
-POM_DEVELOPER_EMAIL=andrew@git.live
+# Developer information
+POM_DEVELOPER_ID=trykovyura
+POM_DEVELOPER_NAME=Yuri
+POM_DEVELOPER_EMAIL=help@surprisesantas.com
 
+# Kotlin configuration
 kotlin.native.binary.memoryModel=experimental
 kotlin.native.cacheKind=none
-
 kotlin.js.compiler=ir
 kotlin.mpp.enableCInteropCommonization=true
 
-signing.keyId=""
-signing.password=""
-
-gpr.user=yet300
+# GitHub Packages configuration
+gpr.user=trykovyura
 gpr.key=todo
 
-MODULE_PACKAGE_NAME=dev.gitlive
+# Signing configuration (optional for GitHub Packages, required for Maven Central)
+# These should be set via environment variables or command line
+# signing.keyId=your-key-id
+# signing.password=your-key-password
+# signing.secretKeyRingFile=path/to/secring.gpg
+
+# Module information
+MODULE_PACKAGE_NAME=dev.yet300
 MODULE_VERSION_NUMBER=0.1.2
 MODULE_NAME=appauth-kotlin
 
-OPEN_SOURCE_REPO=https://oss.sonatype.org/service/local/staging/deploy/maven2/
+# Publishing configuration
 PUBLISH_NAME=appauth-kotlin
-PUBLISH_DESCRIPTION=A Kotlin first OAuth Library
-PUBLISH_URL=https://github.com/GitLiveApp/AppAuth-Kotlin
+PUBLISH_DESCRIPTION=A Kotlin-first OAuth Library for multiplatform projects
+PUBLISH_URL=https://github.com/yet300/AppAuth-Kotlin
 
-PUBLISH_SCM_URL=https://github.com/GitLiveApp/AppAuth-Kotlin
-PUBLISH_SCM_CONNECTION=scm:git://git@github.com:GitLiveApp/AppAuth-Kotlin.git
-PUBLISH_SCM_DEVELOPERCONNECTION=scm:git:ssh://git@github.com:GitLiveApp/AppAuth-Kotlin.git
+# SCM information
+PUBLISH_SCM_URL=https://github.com/yet300/AppAuth-Kotlin
+PUBLISH_SCM_CONNECTION=scm:git://git@github.com:yet300/AppAuth-Kotlin.git
+PUBLISH_SCM_DEVELOPERCONNECTION=scm:git:ssh://git@github.com:yet300/AppAuth-Kotlin.git
