+++ src/androidUnitTest/kotlin/dev/yet300/appauth/commonTest.android.kt
@@ -1,0 +1,14 @@
+package dev.yet300.appauth
+
+import kotlinx.coroutines.CoroutineScope
+
+actual val context: Any
+    get() = TODO("Not yet implemented")
+
+actual fun simulateSignIn() {
+    TODO("Not yet implemented")
+}
+
+actual suspend fun CoroutineScope.withAuthorizationService(action: suspend (AuthorizationService) -> Unit) {
+    TODO("Not yet implemented")
+}
