+++ src/iosTest/kotlin/dev/yet300/appauth/iosTest.kt
@@ -1,0 +1,12 @@
+package dev.yet300.appauth
+
+import kotlinx.coroutines.CoroutineScope
+
+actual val context: Any
+    get() = TODO("context not yet implemented")
+
+actual fun simulateSignIn() {
+}
+
+actual suspend fun CoroutineScope.withAuthorizationService(action: suspend (service: AuthorizationService) -> Unit) {
+}
