--- src/jsTest/kotlin/dev/gitlive/appauth/jsTest.kt
@@ -1,12 +1,0 @@
-package dev.gitlive.appauth
-
-import kotlinx.coroutines.CoroutineScope
-
-actual val context: Any
-    get() = TODO("Not yet implemented")
-
-actual fun simulateSignIn() {
-}
-
-actual suspend fun CoroutineScope.withAuthorizationService(action: suspend (service: AuthorizationService) -> Unit) {
-}
